Adding more error handling and split agent factory method into separate file
